Change network detection to Web3Context
The wrong network alert is shown from Web3Context and a rightNetwork flag is set to be able to know (using the context) if the user has the right network selected.

Use readContracts from context to be able to show dev profiles even if the user is not on the right network.